+++ src/main/java/ru/mt/domain/Reservation.java
@@ -5,6 +5,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import ru.mt.utils.TimeUtils;
 
+import java.math.BigDecimal;
+
 /**
  * Запись о резервировании суммы денег на счете
  */
@@ -15,7 +17,7 @@ public class Reservation {
 
     private final String accountId;
     private final String transactionId;
-    private final double amount;
+    private final BigDecimal amount;
     @Setter
     private ReservationStatus status;
 }
