+++ src/main/java/ru/mt/domain/Transaction.java
@@ -3,6 +3,7 @@ package ru.mt.domain;
 import lombok.*;
 import ru.mt.utils.TimeUtils;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -16,10 +17,10 @@ public class Transaction {
 
     private final String accountIdFrom;
     private final String accountIdTo;
-    private final double amount;
+    private final BigDecimal amount;
     private final TransactionStatus status;
 
-    public Transaction(String accountIdFrom, String accountIdTo, double amount, TransactionStatus status) {
+    public Transaction(String accountIdFrom, String accountIdTo, BigDecimal amount, TransactionStatus status) {
         this.accountIdFrom = accountIdFrom;
         this.accountIdTo = accountIdTo;
         this.amount = amount;
